+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Spawn multiple commands in parallel with `--threads`.
 - Add options for command output handling:
   - `--no-header`
   - `--result` with `always`, `never` and `non-zero`